pull the Forestage component out of the Menu component
